Refactor asset creation to use configuration
